camera system works in theory

+++ src/scomponents/io/inputs.h
@@ -16,6 +16,6 @@ namespace scomp {
 		std::array<bool, InputAction::_ACTION_MAX> actionState;
 		glm::vec2 mousePos = { 0, 0 };
 		glm::vec2 delta = { 0, 0 };
-		unsigned int wheelDelta = 0;
+		short wheelDelta = 0;
 	};
 }
